Major restructure of data structures.
