+++ pyaestro/core/datastructures/abstracts/__init__.py
@@ -1,0 +1,3 @@
+from .graphs import Graph
+
+__all__ = ("Graph")
