+++ pyaestro/core/datastructures/graphs/__init__.py
@@ -1,79 +1,30 @@
-from typing import Hashable, Iterable, Tuple
+import functools
+from typing import Dict, Hashable, Type
 
 from ..abstracts import Graph
+from .algorithms import detect_cycles
 
 
-class AdjacencyGraph(Graph):
-    """An adjacency list implementation of the abstract Graph API."""
-    def __init__(self):
-        self._adj_table = {}
-        super().__init__()
+class AcyclicGraph(Graph):
+    """An acyclic variant of the Graph abstract class."""
 
-    def __setitem__(self, key: Hashable, value: object) -> None:
-        super().__setitem__(key, value)
-        if key not in self._adj_table:
-            self._adj_table[key] = set()
+    def _cycle_check(function):
+        @functools.wraps(function)
+        def cycle_check_wrapper(*args, **kwargs):
+            ret_value = function(*args, **kwargs)
+            if detect_cycles(args[0]):
+                raise Exception("Cycle detected")
+            return ret_value
+        return cycle_check_wrapper
 
-    def __delitem__(self, key: Hashable) -> None:
-        try:
-            super().__delitem__(key)
-            del self._adj_table[key]
-        except KeyError as key_error:
-            raise KeyError(f"Key '{key_error.args[0]}' not found in graph.")
-
-    def edges(self) -> Iterable[Tuple[Hashable]]:
-        for src, adj_list in self._adj_table.items():
-            for dest in adj_list:
-                yield src, dest
-
-    def get_neighbors(self, node: Hashable) -> Iterable[Hashable]:
-        try:
-            for dest in self._adj_table[node]:
-                yield dest
-        except KeyError as key_error:
-            raise KeyError(f"Key '{key_error.args[0]}' not found in graph.")
-
-    def delete_edges(self, key: Hashable) -> None:
-        try:
-            self._adj_table[key].discard(key)
-            for neighbor in self._adj_table[key]:
-                self._adj_table[neighbor].remove(key)
-            self._adj_table[key].clear()
-        except KeyError as key_error:
-            raise KeyError(f"Key '{key_error.args[0]}' not found in graph.")
-
+    @_cycle_check
     def add_edge(self, a: Hashable, b: Hashable) -> None:
-        """Add a bidirectional edge between node 'a' and node 'b' to the graph.
-
-        Args:
-            a (Hashable): Key identifying side 'a' of an edge.
-            b (Hashable): Key identifying side 'b' of an edge.
-
-        Raises:
-            KeyError: Raised when either node 'a' or node 'b'
-            do not exist in the graph.
-        """
-        try:
-            # Add each edge
-            self._adj_table[a].add(b)
-            self._adj_table[b].add(a)
-        except KeyError as key_error:
-            raise KeyError(f"Key '{key_error.args[0]}' not found in graph.")
-
-    def remove_edge(self, a: Hashable, b: Hashable) -> None:
-        """Remove a edge between node 'a' and node 'b' from the graph.
-
-        Args:
-            a (Hashable): Key identifying side 'a' of an edge.
-            b (Hashable): Key identifying side 'b' of an edge.
-
-        Raises:
-            KeyError: Raised when either node 'a' or node 'b'
-            do not exist in the graph.
-        """
-        try:
-            # Bidirectional graph will cause a key error if not checked.
-            self._adj_table[a].remove(b)
-            self._adj_table[b].discard(a)
-        except KeyError as key_error:
-            raise KeyError(f"Key '{key_error.args[0]}' not found in graph.")
+        super().add_edge(a, b)
+        
+    @_cycle_check
+    @classmethod
+    def from_specification(
+        cls,
+        specification: Dict[Hashable, Dict[Hashable, object]]
+    ) -> Type[Graph]:
+        return super().from_specification(specification)
